Improve Ownership Transfer System
Fixed bugs in Ownership Volume causing it to not identify its owner. Ownership Volumes now track their managed Ownership Managers. Changes to the current owner are propagated to all contained ownership managers.

Added a sphere collider to the players head (it may help with camera clipping) but its main purpose is to provide a collider to represent the player so they have a collider which can collider with an ownership volume to give them ownership.

Fixed bug in NetworkRigidbody(2D) that was preventing it from ever becoming unkinematic.