Slot Upgrade
Slots on the PlayerAvatar are now created with a dictionary which maps slot name to pose reference. Thus an avatar can be configured with any arbitrary slots.
SyncPoses have been upgraded with a custom UI (very similar in appearance to the old UI) with the slot dropdown list automatically filled with the slot names from the associated player avatar dictionary.