FishyVoice
Packaged together all of the voice related code into a small package (we may distribute this separately at some point). The files have been added to a namespace and renamed for clarity.
Simplified the voice GUI: it is now a simple button that works similarly to the server and client buttons in the default UI provided by FishNet.